Revert "Add A-MoreOOP"
This reverts commit f2831a0153fed7a0b372eacb3c1f33c2ecaa5550.
