--- src/main/java/Deadline.java
@@ -1,21 +1,0 @@
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-public class Deadline extends Task {
-    protected LocalDateTime by;
-
-    public Deadline(String description, String by) {
-        super(description);
-        this.by = LocalDateTime.parse(by, DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm"));
-    }
-
-    @Override
-    public String toFileString() {
-        return String.format("D | %d | %s | %s", isDone ? 1 : 0, description, by.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm")));
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + by.format(DateTimeFormatter.ofPattern("MMM d yyyy, h:mma")) + ")";
-    }
-}
