--- src/main/java/JardException.java
@@ -1,5 +1,0 @@
-public class JardException extends RuntimeException {
-    public JardException(String message) {
-        super(message);
-    }
-}
