--- src/main/java/TaskList.java
@@ -1,34 +1,0 @@
-import java.util.ArrayList;
-import java.util.List;
-
-public class TaskList {
-    private final List<Task> tasks;
-
-    public TaskList() {
-        this.tasks = new ArrayList<>();
-    }
-
-    public TaskList(List<Task> tasks) {
-        this.tasks = tasks;
-    }
-
-    public List<Task> getTasks() {
-        return tasks;
-    }
-
-    public void addTask(Task task) {
-        tasks.add(task);
-    }
-
-    public Task deleteTask(int index) {
-        return tasks.remove(index);
-    }
-
-    public Task getTask(int index) {
-        return tasks.get(index);
-    }
-
-    public int size() {
-        return tasks.size();
-    }
-}
