--- src/main/java/Todo.java
@@ -1,15 +1,0 @@
-public class Todo extends Task {
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toFileString() {
-        return String.format("T | %d | %s", isDone ? 1 : 0, description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
