--- src/main/java/Ui.java
@@ -1,86 +1,0 @@
-import java.util.Scanner;
-import java.util.List;
-
-public class Ui {
-    private Scanner scanner;
-
-    public Ui() {
-        scanner = new Scanner(System.in);
-    }
-
-    public void showWelcome() {
-        String logo = " ____        _        \n"
-                + "|  _ \\ _   _| | _____ \n"
-                + "| | | | | | | |/ / _ \\\n"
-                + "| |_| | |_| |   <  __/\n"
-                + "|____/ \\__,_|_|\\_\\___|\n";
-        System.out.println("Hello from\n" + logo);
-        System.out.println("____________________________________________________________");
-        System.out.println(" Hello! I'm Jard.");
-        System.out.println(" What can I do for you?");
-        System.out.println("____________________________________________________________");
-    }
-
-    public void showBye() {
-        System.out.println("____________________________________________________________");
-        System.out.println("Bye. Hope to see you again soon!");
-        System.out.println("____________________________________________________________");
-    }
-
-    public void showTaskList(List<Task> tasks) {
-        if (tasks.isEmpty()) {
-            System.out.println("Nothing in the list!");
-        } else {
-            System.out.println("____________________________________________________________");
-            System.out.println("Here are the tasks in your list:");
-            for (int i = 0; i < tasks.size(); i++) {
-                System.out.println(" " + (i + 1) + ". " + tasks.get(i));
-            }
-            System.out.println("____________________________________________________________");
-        }
-    }
-
-    public void showAddTask(Task task, int size) {
-        System.out.println("____________________________________________________________");
-        System.out.println("Got it. I've added this task:");
-        System.out.println("   " + task);
-        System.out.println("Now you have " + size + " tasks in the list.");
-        System.out.println("____________________________________________________________");
-    }
-
-    public void showMarkTask(Task task) {
-        System.out.println("____________________________________________________________");
-        System.out.println("Nice! I've marked this task as done:");
-        System.out.println("   " + task);
-        System.out.println("____________________________________________________________");
-    }
-
-    public void showUnmarkTask(Task task) {
-        System.out.println("____________________________________________________________");
-        System.out.println("OK, I've marked this task as not done yet:");
-        System.out.println("   " + task);
-        System.out.println("____________________________________________________________");
-    }
-
-    public void showDeleteTask(Task task, int size) {
-        System.out.println("____________________________________________________________");
-        System.out.println(" Noted. I've removed this task:");
-        System.out.println("   " + task);
-        System.out.println(" Now you have " + size + " tasks in the list.");
-        System.out.println("____________________________________________________________");
-    }
-
-    public void showError(String message) {
-        System.out.println("____________________________________________________________");
-        System.out.println(" Error: Jard! " + message);
-        System.out.println("____________________________________________________________");
-    }
-
-    public String readCommand() {
-        return scanner.nextLine().trim();
-    }
-
-    public void close() {
-        scanner.close();
-    }
-}
